Implement inventory ESP for players
This script creates an ESP (Extra Sensory Perception) feature that displays the inventory of other players under their feet, showing items in their Backpack and what they are currently holding. It includes functions for updating the display when items are added or removed, as well as handling player connections and disconnections.